sdkman function moved to unix
